Fix corner cases on neighbor search.
This bugfix addresses a numerical rounding issue
that occurs when integer multiples of the sub-cell
z-size are "close" to the search cutoff radius.
In that case, R2z could become "-0.0", which could cause
a domain error in sqrt.  A new test eliminates this
possibility.  Similarly for R2y, a nearly zero distance
will now advance the search more conservatively (to j++
instead of k++).